Generate barcode in an AsyncTask
This change moves the generation of the barcode into its
own async task. In addition, the size of the ImageView is
used to determine the barcode size to use.

There will be cases when the size of the ImageView will not
be known when the barcode generation starts. This will be resolved
in a future commit.