Allow user to enter a barcode manually
If a user is unable to scan a barcode, this commit allows a
user to enter is manually.

If the user selects to Enter Card instead of Capture Card,
the user may enter the card's id. As it may not be known which barcode
format the user expects, and the user may not know what barcode
type is what, all barcode types are generated from the user
input. Those that are valid are displayed to the user. The user
may then select the barcode image which matches what the user wants.

Italian translations provided by Michael Moroni (Airon90)
Dutch translations provided by PanderMusubi