Update first start wizard images with new logo and colors
Note that the second slide has been changed. Originally there was
text over the image with an arrow. This text could not be translated
and was part of the instruction. The arrow is really not needed,
so it has been removed with the overlay text.

Also, the originals folder has been removed. The screenshots which
were used to create the wizard images are in the metadata folder.