Reduce space used in header for card view
The landscape view left little room for the barcode and card ID
because the header took up a lot of space. Further, on smaller
phones it would take up a large portion of the screen.

This change reduces the header from 224dp to two action bar sizes
plus the size of the textview for the store name.