Remove rotation support
With the release of targetSdk 36, the Android platform will no longer
always apply rotation. To lower the chance of user confusion, this
support is being removed completely.

I know some of you really love this, but I can't do anything about this.
This is Google's decision and my opinion doesn't matter.