Fix crash on missing header colors
An off-by-one error caused invalid colour info to sometimes be generated
if no header colour was set. Under normal conditions, a header colour
should always be set, but due to some bugs in the past in some cases
they may not be set.

Sadly this does change the card colours for cards which are not properly
initialized. But that's better than a crash at least.